chore: separate parsleyjs dependency
