+++ src/form-validation-parsley.html
@@ -120,6 +120,7 @@
 {% endblock %}
 
 {% block js %}
-<script src="assets/js/extensions/parsley.init.js"></script>
-<script src="assets/js/extensions/parsley.custom.js"></script>
+<script src="assets/extensions/jquery/jquery.min.js"></script>
+<script src="assets/extensions/parsleyjs/parsley.min.js"></script>
+<script src="assets/js/pages/parsley.js"></script>
 {% endblock %}
